viam_modules/hello-world/requirements.txt

viam-sdk==0.45.2

typing-extensions
Pillow
